+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 ## current master
 * add supportedInterfaces to request builder
+* add missing feature `withSessionAttributes` (#3)
 
 ## 2.0.1
 * adding more supported types to attributes check